Moving lib and common to src
