+++ src/common/exceptions.py
